// src/app/models/common/Container.ts
export class Container{
    ClientData: any;
    ContainerID: string;
    CategoryName: string;
    Location: string;
    ClientId: string;
    ContainerCycle: number;
    ContainerCapacity: number;
    Status: number;
    StatusText: string;
    Quantity: number;
    Properties: string;
    ParentContainer: any;
    ContainerType: string;
    InvAccount:string;
}