// src/app/enums/common.enum.ts
export enum CommonEnum {
  add = 'Add',
  save = 'Save',
  yes = 'Y',
  no = 'N',
  lblYes = 'YES',
  lblNo = 'NO',
  update = 'Update',
  inprocess = 'INPROCESS',
  pass = 'PASS',
  automationQueue = 'Automation Queue',
  samplingDetails = 'Sampling Details',
  routeDetails = 'Route Details',
  containerDetails = 'Container Details',
  processQueue = 'Process Queue',
  batchDetails = 'Batch Details',
  deviceDetails = 'Device Details',
  route = 'ROUTE',
  device = 'Device',
  attributeTypes = 'ATTRIBUTETYPES',
  attributes = 'ATTRIBUTES',
  dispositionEngineAttribute = 'DISPOSITIONENGINEATTR',
  deviceCatalogAttr='DEVICECATALOGATTR',
  autoFailLater = 'AUTOFAILATTR',
  programName = 'PROGRAMNAME',
  programIndicator = 'PROGRAMINDICATOR',
  disposition = 'DISPOSITION',
  conditionCode = 'CONDITIONCODE',
  oemCode='GetOEMcodes',
  lookypType = 'LOOKUPTYPE',
  skuLookupAttr='DEVICECATXREFATTR',
  Item='ITEM_ID',
  SKU='SKU',
  us = 'USA',
  and = 'AND',
  or = 'OR',
  controlType = 'TEXTBOX',
  in = 'in',
  notOperator = '!',
  boolean = 'Boolean',
  ID = 'ID',
  Id = 'Id',
  TEXT = 'TEXT',
  Text = 'Text',
  attributeID = 'ATTRIBUTEID',
  attributeType = 'ATTR_TYPE',
  attributeProperty='ATTR_PROPERTY',
  groupType='TYPE',
  gengroupID='Gen Group ID',
  gengroupValue='DEFAULT',
  ruleSetup = 'RuleSetUp',
  openParanthesis = '(',
  closeParanthesis = ')',
  reset = 'RESET',
  clear = 'CLEAR',
  MOVEBATCH = 'Move Batch',
  CLOSECONTAINER = 'Close Container',
  CLOSEBATCH = 'Close Batch',
  new = 'NEW',
  notCompleted = 'Not Completed',
  completed = 'Completed',
  box = 'BOX',
  videoinput = 'videoinput',
  error = 'ERROR',
  checkAll = 'checkAll',
  retry = 'Retry',
  autoClear = 'Clear',
  manual = 'Manual',
  reassign = 'Reassign',
  override = "OVERRIDE",
  remove = 'REMOVE',
  revive = 'REVIVE',
  itemType = 'ITEMTYPE',
  oem='OEM',
  carrierCode = 'CARRIERCODE',
  modelOptions = 'MODELOPTIONS',
  catalogType = 'CATALOGTYPE',
  start = 'Start',
  stop = 'Stop',
  TotalCount ='Total Count',
  InProgress ='In-Progress',
  Failed ='Failed',
  DeviceCount ='Device Count'
}